JSON.parse and JSON.stringify built, from scratch
JSON.parse came together quick, because I understand that the often
disparaged eval function did something pretty close to what I wanted,
i.e. looking at a string and making it into JS. The stringily was
straightforward but because of escape characters and the various
different data types in JS it took a while to make this but it looks
good. As before, console logging out the result and the expected really
helped me see the gaps between where I was and where I wanted to go.